+++ examples/server_hardware_type.rb
@@ -1,7 +1,3 @@
-#
-# Cookbook Name:: oneview_test
-# Recipe:: storage_pool_add_with_name
-#
 # (c) Copyright 2016 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,9 +8,23 @@
 # under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
-#
 
-oneview_storage_pool 'StoragePool2' do
-  client node['oneview_test']['client']
-  storage_system_name 'StorageSystem1'
+my_client = {
+  url: '',
+  user: '',
+  password: ''
+}
+
+# Example: Update server hardware type properties
+oneview_server_hardware_type 'DL360 Gen9 1' do
+  client my_client
+  data(
+    description: 'Server hardware type description'
+  )
+end
+
+# Example: Remove server hardware type
+oneview_server_hardware_type 'DL360 Gen9 1' do
+  client my_client
+  action :remove
 end