+++ examples/switch.rb
@@ -15,39 +15,17 @@ my_client = {
   password: ''
 }
 
-# It will not do anything if no action is selected
-oneview_interconnect 'Encl1, interconnect 1' do
+# Example: No action is executed.
+# In a resource that has only one action and no action is specified in the block
+# Chef executes this one. To prevent Chef from removing a switch as the standard action we created
+# the none action
+oneview_switch 'Switch1' do
   client my_client
 end
 
-oneview_interconnect 'Encl1, interconnect 1' do
-  client my_client
-  action :reset_port_protection
-end
 
-oneview_interconnect 'Encl1, interconnect 1' do
+# Example: removes switch
+oneview_switch 'Switch1' do
   client my_client
-  uid_light_state 'On'
-  action :set_uid_light
-end
-
-oneview_interconnect 'Encl1, interconnect 1' do
-  client my_client
-  power_state 'On'
-  action :set_power_state
-end
-
-oneview_interconnect 'Encl1, interconnect 1' do
-  client my_client
-  port_options(
-    name: 'X4',
-    portName: 'X4',
-    enabled: true
-  )
-  action :update_port
-end
-
-oneview_interconnect 'Encl1, interconnect 1' do
-  client my_client
-  action :reset
+  action :remove
 end