+++ resources/server_hardware_type.rb
@@ -1,7 +1,3 @@
-#
-# Cookbook Name:: oneview_test
-# Recipe:: firmware_bundle_add
-#
 # (c) Copyright 2016 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,9 +8,22 @@
 # under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
-#
 
-oneview_firmware_bundle 'upload cp027376.exe' do
-  client node['oneview_test']['client']
-  file_path '/path/to/cp027376.exe'
+OneviewCookbook::ResourceBaseProperties.load(self)
+
+default_action :edit
+
+action_class do
+  include OneviewCookbook::Helper
+  include OneviewCookbook::ResourceBase
+end
+
+action :edit do
+  item = load_resource
+  return add_or_edit(item) if item.exists?
+  raise "Resource not found: #{resource_name} '#{item['name']}'"
+end
+
+action :remove do
+  remove
 end