+++ spec/fixtures/cookbooks/oneview_test/recipes/firmware_create_custom_spp_invalid_hotfix.rb
@@ -1,3 +1,7 @@
+#
+# Cookbook Name:: oneview_test
+# Recipe:: firmware_create_custom_spp_invalid_hotfix
+#
 # (c) Copyright 2016 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -8,29 +12,10 @@
 # under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
+#
 
-my_client = {
-  url: '',
-  user: '',
-  password: ''
-}
-
-# Example: Adds storage pool if it is not added in HPE OneView using
-# the storage system name
-oneview_storage_pool 'CPG_FC-AO' do
-  client my_client
-  storage_system 'ThreePAR7200-8147'
-end
-
-# Example: Adds storage pool if it is not added in HPE OneView using
-# the storage system hostname
-oneview_storage_pool 'CPG_FC-AO' do
-  client my_client
-  storage_system '172.XX.XX.XX'
-end
-
-# Example: Removes storage pool from HPE OneView
-oneview_storage_pool 'CPG_FC-AO' do
-  client my_client
-  action :remove
+oneview_firmware 'CustomSPP1' do
+  client node['oneview_test']['client']
+  spp_name 'Service Pack for ProLiant'
+  action :create_custom_spp
 end