+++ spec/fixtures/cookbooks/oneview_test/recipes/server_hardware_type_remove.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: oneview_test
-# Recipe:: storage_pool_add_with_ip
+# Recipe:: server_hardware_type_remove
 #
 # (c) Copyright 2016 Hewlett Packard Enterprise Development LP
 #
@@ -14,7 +14,7 @@
 # specific language governing permissions and limitations under the License.
 #
 
-oneview_storage_pool 'StoragePool1' do
+oneview_server_hardware_type 'ServerHardwareType1' do
   client node['oneview_test']['client']
-  storage_system_ip '10.1.1.1'
+  action :remove
 end