+++ spec/fixtures/cookbooks/oneview_test/recipes/storage_system_add_if_missing.rb
@@ -1,3 +1,7 @@
+#
+# Cookbook Name:: oneview_test
+# Recipe:: storage_system_add_if_missing
+#
 # (c) Copyright 2016 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -8,29 +12,17 @@
 # under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
+#
 
-my_client = {
-  url: '',
-  user: '',
-  password: ''
-}
-
-# Example: Adds storage pool if it is not added in HPE OneView using
-# the storage system name
-oneview_storage_pool 'CPG_FC-AO' do
-  client my_client
-  storage_system 'ThreePAR7200-8147'
-end
-
-# Example: Adds storage pool if it is not added in HPE OneView using
-# the storage system hostname
-oneview_storage_pool 'CPG_FC-AO' do
-  client my_client
-  storage_system '172.XX.XX.XX'
-end
-
-# Example: Removes storage pool from HPE OneView
-oneview_storage_pool 'CPG_FC-AO' do
-  client my_client
-  action :remove
+oneview_storage_system 'StorageSystem1' do
+  client node['oneview_test']['client']
+  data(
+    credentials: {
+      ip_hostname: '127.0.0.1',
+      username: 'username',
+      password: 'password'
+    },
+    managedDomain: 'TestDomain'
+  )
+  action :add_if_missing
 end