+++ CHANGELOG.md
@@ -15,6 +15,12 @@
   - Added support to Logical switch group actions
   - Added support to Logical switch actions
   - Added support to Firmware bundle actions
+  - Added support to Server hardware type actions
+  - Added support to Power device actions
+  - Changed Storage pool action `:add` to `:add_if_missing`
+  - Added support to Storage system `:edit_credentials` and `:add_if_missing`
+  - Added support to Switch actions
+  - Added support to Firmware driver actions and integrated it with firmware bundles in a resource called Firmware
   - Added diff output for updating resources
   - Added support to Manage san actions